+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 
 ## Unreleased
 
+### Added
+- (`sklearn`) `keep_selected_variables_only` parameter to the predictors (`KhiopsClassifier` and `KhiopsRegressor`)
+
 ### Changed
 - (`core`) Rename `variable_part_dimensions` to `inner_variable_dimensions` in Coclustering results.